Added Analyzers, GL context and window creation.
Started to move data analysis from the timestep constructors to
factories.
Added analyzerinterface for that reason and started the
implementations of gaussiananalyzer. (Remove analysis from
timestep constructor, when gaussiananalyzer is ready)
Reconsider friend class situation of the analyzer classes and
timestep!! There has to be a way to register/inherit access to
timesteps members.

Added Ensemble as pool/manager class for all analyzed timesteps.
Should be constructed by a path to the ensemble and then uses an
analyzer to fill a buffer of analyzed timesteps.

Added OpenGL context and window creation.